Fix for unsupport custom SSL/TLS certificate handling due to libcurl and it's SSL backend.
This works on certain setups but the best change is one that is the most universal.

System.AggregateException: One or more errors occurred. (The handler does not support custom handling of certificates with this combination of libcurl (7.58.0) and its SSL backend ("OpenSSL/1.1.0g").) ---> System.PlatformNotSupportedException: The handler does not support custom handling of certificates with this combination of libcurl (7.58.0) and its SSL backend ("OpenSSL/1.1.0g").